Extract durable call session coordination
