+++ golem-worker-executor/src/durable_host/clocks/monotonic_clock.rs
@@ -14,7 +14,7 @@
 
 use wasmtime::component::Resource;
 
-use crate::durable_host::concurrent::{CallHandle, NotCancellable};
+use crate::durable_host::concurrent::{DurableCallSession, NotCancellable};
 use crate::durable_host::{DurabilityHost, DurableWorkerCtx};
 use crate::services::HasWorker;
 use crate::services::oplog::CommitLevel;
@@ -28,15 +28,16 @@ use wasmtime_wasi::p2::bindings::clocks::monotonic_clock::{Duration, Host, Insta
 
 impl<Ctx: WorkerCtx> Host for DurableWorkerCtx<Ctx> {
     async fn now(&mut self) -> wasmtime::Result<Instant> {
-        // `now()` is a re-executable `ReadLocal`, so it uses the `CallHandle::run` combinator: the
+        // `now()` is a re-executable `ReadLocal`, so it uses the `DurableCallSession::run` combinator: the
         // live clock read is supplied as the action and is run on the live path or re-run if replay
         // finds the `Start` without its `End`; a committed `End` replays without touching the clock.
-        let handle = CallHandle::<host_functions::MonotonicClockNow, NotCancellable>::start(
-            self,
-            HostRequestNoInput {},
-            DurableFunctionType::ReadLocal,
-        )
-        .await?;
+        let handle =
+            DurableCallSession::<host_functions::MonotonicClockNow, NotCancellable>::start(
+                self,
+                HostRequestNoInput {},
+                DurableFunctionType::ReadLocal,
+            )
+            .await?;
 
         let result = handle
             .run(self, async |ctx| -> wasmtime::Result<_> {
@@ -50,12 +51,13 @@ impl<Ctx: WorkerCtx> Host for DurableWorkerCtx<Ctx> {
     }
 
     async fn resolution(&mut self) -> wasmtime::Result<Instant> {
-        let handle = CallHandle::<host_functions::MonotonicClockResolution, NotCancellable>::start(
-            self,
-            HostRequestNoInput {},
-            DurableFunctionType::ReadLocal,
-        )
-        .await?;
+        let handle =
+            DurableCallSession::<host_functions::MonotonicClockResolution, NotCancellable>::start(
+                self,
+                HostRequestNoInput {},
+                DurableFunctionType::ReadLocal,
+            )
+            .await?;
 
         let result = handle
             .run(self, async |ctx| -> wasmtime::Result<_> {
@@ -80,13 +82,15 @@ impl<Ctx: WorkerCtx> Host for DurableWorkerCtx<Ctx> {
         &mut self,
         duration_in_nanos: Duration,
     ) -> wasmtime::Result<Resource<Pollable>> {
-        let handle =
-            CallHandle::<host_functions::MonotonicClockSubscribeDuration, NotCancellable>::start(
-                self,
-                HostRequestMonotonicClockDuration { duration_in_nanos },
-                DurableFunctionType::ReadLocal,
-            )
-            .await?;
+        let handle = DurableCallSession::<
+            host_functions::MonotonicClockSubscribeDuration,
+            NotCancellable,
+        >::start(
+            self,
+            HostRequestMonotonicClockDuration { duration_in_nanos },
+            DurableFunctionType::ReadLocal,
+        )
+        .await?;
 
         let now = handle
             .run(self, async |ctx| -> wasmtime::Result<_> {
