+++ golem-worker-executor/src/durable_host/io/streams.rs
@@ -15,7 +15,7 @@
 use wasmtime::component::Resource;
 use wasmtime_wasi::StreamError;
 
-use crate::durable_host::concurrent::{CallHandle, CallReplayOutcome, NotCancellable};
+use crate::durable_host::concurrent::{CallReplayOutcome, DurableCallSession, NotCancellable};
 use crate::durable_host::durability::HostFailureKind;
 use crate::durable_host::http::{continue_http_request, end_http_request};
 use crate::durable_host::io::{ManagedStdErr, ManagedStdOut};
@@ -61,12 +61,13 @@ impl<Ctx: WorkerCtx> HostInputStream for DurableWorkerCtx<Ctx> {
             let begin_idx = get_http_request_begin_idx(self, handle)?;
 
             let request = get_http_stream_request(self, handle)?;
-            let mut call = CallHandle::<HttpTypesIncomingBodyStreamRead, NotCancellable>::start(
-                self,
-                request,
-                DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
-            )
-            .await?;
+            let mut call =
+                DurableCallSession::<HttpTypesIncomingBodyStreamRead, NotCancellable>::start(
+                    self,
+                    request,
+                    DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
+                )
+                .await?;
 
             let result = if call.is_live() {
                 let first_try = HostInputStream::read(self.table(), self_, len).await;
@@ -128,7 +129,7 @@ impl<Ctx: WorkerCtx> HostInputStream for DurableWorkerCtx<Ctx> {
             // chunk is recorded (the bytes themselves are not needed, as the
             // initial file system is restored to the same contents for replay),
             // and replay re-reads exactly that many bytes from the file.
-            let call = CallHandle::<FilesystemInputStreamRead, NotCancellable>::start(
+            let call = DurableCallSession::<FilesystemInputStreamRead, NotCancellable>::start(
                 self,
                 HostRequestNoInput {},
                 DurableFunctionType::ReadLocal,
@@ -188,13 +189,15 @@ impl<Ctx: WorkerCtx> HostInputStream for DurableWorkerCtx<Ctx> {
             let begin_idx = get_http_request_begin_idx(self, handle)?;
 
             let request = get_http_stream_request(self, handle)?;
-            let mut call =
-                CallHandle::<HttpTypesIncomingBodyStreamBlockingRead, NotCancellable>::start(
-                    self,
-                    request,
-                    DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
-                )
-                .await?;
+            let mut call = DurableCallSession::<
+                HttpTypesIncomingBodyStreamBlockingRead,
+                NotCancellable,
+            >::start(
+                self,
+                request,
+                DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
+            )
+            .await?;
             let result = if call.is_live() {
                 let first_try = HostInputStream::blocking_read(self.table(), self_, len).await;
 
@@ -262,12 +265,13 @@ impl<Ctx: WorkerCtx> HostInputStream for DurableWorkerCtx<Ctx> {
             let begin_idx = get_http_request_begin_idx(self, handle)?;
 
             let request = get_http_stream_request(self, handle)?;
-            let mut call = CallHandle::<HttpTypesIncomingBodyStreamSkip, NotCancellable>::start(
-                self,
-                request,
-                DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
-            )
-            .await?;
+            let mut call =
+                DurableCallSession::<HttpTypesIncomingBodyStreamSkip, NotCancellable>::start(
+                    self,
+                    request,
+                    DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
+                )
+                .await?;
             let result = if call.is_live() {
                 let result = HostInputStream::skip(self.table(), self_, len).await;
                 call.try_trigger_retry(self, &ignore_closed_error(&result), |_| {
@@ -301,7 +305,7 @@ impl<Ctx: WorkerCtx> HostInputStream for DurableWorkerCtx<Ctx> {
             // on host scheduling, so the skipped length is recorded and replay
             // skips exactly that many bytes from the restored file.
             let handle = self_.rep();
-            let call = CallHandle::<FilesystemInputStreamSkip, NotCancellable>::start(
+            let call = DurableCallSession::<FilesystemInputStreamSkip, NotCancellable>::start(
                 self,
                 HostRequestNoInput {},
                 DurableFunctionType::ReadLocal,
@@ -366,13 +370,15 @@ impl<Ctx: WorkerCtx> HostInputStream for DurableWorkerCtx<Ctx> {
             let begin_idx = get_http_request_begin_idx(self, handle)?;
 
             let request = get_http_stream_request(self, handle)?;
-            let mut call =
-                CallHandle::<HttpTypesIncomingBodyStreamBlockingSkip, NotCancellable>::start(
-                    self,
-                    request,
-                    DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
-                )
-                .await?;
+            let mut call = DurableCallSession::<
+                HttpTypesIncomingBodyStreamBlockingSkip,
+                NotCancellable,
+            >::start(
+                self,
+                request,
+                DurableFunctionType::WriteRemoteBatched(Some(begin_idx)),
+            )
+            .await?;
 
             let result = if call.is_live() {
                 let result = HostInputStream::blocking_skip(self.table(), self_, len).await;
@@ -445,13 +451,14 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
         let rep = self_.rep();
         if is_outgoing_http_body_stream(self, rep) {
             let state = get_http_output_stream_state(self, rep)?;
-            let call = CallHandle::<HttpTypesOutgoingBodyStreamCheckWrite, NotCancellable>::start(
-                self,
-                state.request,
-                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-            )
-            .await
-            .map_err(StreamError::from)?;
+            let call =
+                DurableCallSession::<HttpTypesOutgoingBodyStreamCheckWrite, NotCancellable>::start(
+                    self,
+                    state.request,
+                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+                )
+                .await
+                .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 // If the peer already closed the request body stream after an
@@ -501,12 +508,13 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
                 // recorded non-zero budget is replayed, the real stream is first
                 // driven to readiness so that subsequent re-executed writes are
                 // permitted by its state machine.
-                let call = CallHandle::<FilesystemOutputStreamCheckWrite, NotCancellable>::start(
-                    self,
-                    HostRequestNoInput {},
-                    DurableFunctionType::ReadLocal,
-                )
-                .await?;
+                let call =
+                    DurableCallSession::<FilesystemOutputStreamCheckWrite, NotCancellable>::start(
+                        self,
+                        HostRequestNoInput {},
+                        DurableFunctionType::ReadLocal,
+                    )
+                    .await?;
 
                 if call.is_live() {
                     let result = HostOutputStream::check_write(self.table(), self_).await;
@@ -570,13 +578,14 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
 
         if is_outgoing_http_body_stream(self, rep) {
             let state = get_http_output_stream_state(self, rep)?;
-            let mut call = CallHandle::<HttpTypesOutgoingBodyStreamWrite, NotCancellable>::start(
-                self,
-                state.request,
-                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-            )
-            .await
-            .map_err(StreamError::from)?;
+            let mut call =
+                DurableCallSession::<HttpTypesOutgoingBodyStreamWrite, NotCancellable>::start(
+                    self,
+                    state.request,
+                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+                )
+                .await
+                .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 // Attempt inline retry BEFORE persisting so the retried result is what gets recorded
@@ -668,13 +677,14 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
             // handles backpressure via write_ready() loops, and persist a single
             // combined oplog entry.
             let state = get_http_output_stream_state(self, rep)?;
-            let mut call = CallHandle::<HttpTypesOutgoingBodyStreamWrite, NotCancellable>::start(
-                self,
-                state.request,
-                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-            )
-            .await
-            .map_err(StreamError::from)?;
+            let mut call =
+                DurableCallSession::<HttpTypesOutgoingBodyStreamWrite, NotCancellable>::start(
+                    self,
+                    state.request,
+                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+                )
+                .await
+                .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 let first_try =
@@ -754,13 +764,14 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
         let rep = self_.rep();
         if is_outgoing_http_body_stream(self, rep) {
             let state = get_http_output_stream_state(self, rep)?;
-            let mut call = CallHandle::<HttpTypesOutgoingBodyStreamFlush, NotCancellable>::start(
-                self,
-                state.request,
-                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-            )
-            .await
-            .map_err(StreamError::from)?;
+            let mut call =
+                DurableCallSession::<HttpTypesOutgoingBodyStreamFlush, NotCancellable>::start(
+                    self,
+                    state.request,
+                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+                )
+                .await
+                .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 // Attempt inline retry BEFORE persisting so the retried result is what gets recorded
@@ -807,14 +818,16 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
         let rep = self_.rep();
         if is_outgoing_http_body_stream(self, rep) {
             let state = get_http_output_stream_state(self, rep)?;
-            let mut call =
-                CallHandle::<HttpTypesOutgoingBodyStreamBlockingFlush, NotCancellable>::start(
-                    self,
-                    state.request,
-                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-                )
-                .await
-                .map_err(StreamError::from)?;
+            let mut call = DurableCallSession::<
+                HttpTypesOutgoingBodyStreamBlockingFlush,
+                NotCancellable,
+            >::start(
+                self,
+                state.request,
+                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+            )
+            .await
+            .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 // Attempt inline retry BEFORE persisting so the retried result is what gets recorded
@@ -890,14 +903,16 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
         let rep = self_.rep();
         if is_outgoing_http_body_stream(self, rep) {
             let state = get_http_output_stream_state(self, rep)?;
-            let mut call =
-                CallHandle::<HttpTypesOutgoingBodyStreamWriteZeroes, NotCancellable>::start(
-                    self,
-                    state.request,
-                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-                )
-                .await
-                .map_err(StreamError::from)?;
+            let mut call = DurableCallSession::<
+                HttpTypesOutgoingBodyStreamWriteZeroes,
+                NotCancellable,
+            >::start(
+                self,
+                state.request,
+                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+            )
+            .await
+            .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 // Attempt inline retry BEFORE persisting so the retried result is what gets recorded
@@ -979,14 +994,16 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
             // Instead, call the underlying blocking_write_zeroes_and_flush which properly
             // handles backpressure, and persist a single combined oplog entry.
             let state = get_http_output_stream_state(self, rep)?;
-            let mut call =
-                CallHandle::<HttpTypesOutgoingBodyStreamWriteZeroes, NotCancellable>::start(
-                    self,
-                    state.request,
-                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-                )
-                .await
-                .map_err(StreamError::from)?;
+            let mut call = DurableCallSession::<
+                HttpTypesOutgoingBodyStreamWriteZeroes,
+                NotCancellable,
+            >::start(
+                self,
+                state.request,
+                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+            )
+            .await
+            .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 // For HTTP body streams, Closed is also retryable (hyper consumer
@@ -1059,13 +1076,14 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
                     .has_unreconstructable_body = true;
             }
             let state = get_http_output_stream_state(self, rep)?;
-            let call = CallHandle::<HttpTypesOutgoingBodyStreamSplice, NotCancellable>::start(
-                self,
-                state.request,
-                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-            )
-            .await
-            .map_err(StreamError::from)?;
+            let call =
+                DurableCallSession::<HttpTypesOutgoingBodyStreamSplice, NotCancellable>::start(
+                    self,
+                    state.request,
+                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+                )
+                .await
+                .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 let result = HostOutputStream::splice(self.table(), self_, src, len).await;
@@ -1127,14 +1145,16 @@ impl<Ctx: WorkerCtx> HostOutputStream for DurableWorkerCtx<Ctx> {
                     .has_unreconstructable_body = true;
             }
             let state = get_http_output_stream_state(self, rep)?;
-            let call =
-                CallHandle::<HttpTypesOutgoingBodyStreamBlockingSplice, NotCancellable>::start(
-                    self,
-                    state.request,
-                    DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
-                )
-                .await
-                .map_err(StreamError::from)?;
+            let call = DurableCallSession::<
+                HttpTypesOutgoingBodyStreamBlockingSplice,
+                NotCancellable,
+            >::start(
+                self,
+                state.request,
+                DurableFunctionType::WriteRemoteBatched(Some(state.begin_index)),
+            )
+            .await
+            .map_err(StreamError::from)?;
 
             let result = if call.is_live() {
                 let result = HostOutputStream::blocking_splice(self.table(), self_, src, len).await;
