+++ golem-worker-executor/src/durable_host/keyvalue/types.rs
@@ -29,7 +29,7 @@ use golem_common::model::oplog::{
 };
 
 use crate::durable_host::concurrent::{
-    AccessClaimOptions, CallHandle, CallReplayOutcome, Cancellable,
+    AccessClaimOptions, CallReplayOutcome, Cancellable, DurableCallSession,
     resolve_current_observational_owner,
 };
 use crate::durable_host::durability::CustomInvocationContext;
@@ -164,21 +164,21 @@ where
         accessor: &Accessor<T, HasSelf<DurableWorkerCtx<Ctx>>>,
     ) -> wasmtime::Result<()> {
         let result = async {
-            let mut handle = CallHandle::<
+            let mut handle = DurableCallSession::<
                 P3KeyvalueTypesIncomingValueConsumeAsync,
                 Cancellable,
             >::start_access_with_options(
-                    accessor,
-                    accessor.getter(),
-                    DurableFunctionType::ReadRemote,
-                    AccessClaimOptions {
-                        observational_owner: self.observational_owner,
-                        ..Default::default()
-                    },
-                    async |_| Ok(HostRequestNoInput {}),
-                )
-                .await
-                .map_err(wasmtime::Error::from)?;
+                accessor,
+                accessor.getter(),
+                DurableFunctionType::ReadRemote,
+                AccessClaimOptions {
+                    observational_owner: self.observational_owner,
+                    ..Default::default()
+                },
+                async |_| Ok(HostRequestNoInput {}),
+            )
+            .await
+            .map_err(wasmtime::Error::from)?;
 
             if !handle.is_live() {
                 match handle
@@ -484,7 +484,7 @@ impl OutgoingValueEntry {
 }
 
 pub(crate) struct CacheFillState {
-    pub handle: Option<CallHandle<P3KeyvalueCacheVacancyFill, Cancellable>>,
+    pub handle: Option<DurableCallSession<P3KeyvalueCacheVacancyFill, Cancellable>>,
     pub environment_id: EnvironmentId,
     pub key: String,
     pub _permit: Option<LiveAuthorizationPermit>,
