+++ golem-worker-executor/src/durable_host/websocket/client.rs
@@ -14,7 +14,7 @@
 
 use crate::durable_host::authorization::targets::websocket_target;
 use crate::durable_host::concurrent::{
-    CallHandle, CallReplayOutcome, LeaveIncompleteOnDrop, NotCancellable,
+    CallReplayOutcome, DurableCallSession, LeaveIncompleteOnDrop, NotCancellable,
 };
 use crate::durable_host::{DurabilityHost, DurableWorkerCtx};
 use crate::preview2::golem::websocket::client::{
@@ -105,11 +105,12 @@ impl<Ctx: WorkerCtx> HostWebsocketConnection for DurableWorkerCtx<Ctx> {
     ) -> anyhow::Result<Result<Resource<WebSocketConnectionEntry>, Error>> {
         self.observe_function_call("golem:websocket/client", "connect");
 
-        let begun = CallHandle::<host_functions::WebsocketClientConnect, NotCancellable>::begin(
-            self,
-            DurableFunctionType::WriteRemote,
-        )
-        .await?;
+        let begun =
+            DurableCallSession::<host_functions::WebsocketClientConnect, NotCancellable>::begin(
+                self,
+                DurableFunctionType::WriteRemote,
+            )
+            .await?;
 
         let mut call = if begun.is_live() {
             let denied = match websocket_target(&url) {
@@ -241,14 +242,15 @@ impl<Ctx: WorkerCtx> HostWebsocketConnection for DurableWorkerCtx<Ctx> {
     ) -> anyhow::Result<Result<(), Error>> {
         self.observe_function_call("golem:websocket/client", "send");
 
-        let mut call = CallHandle::<host_functions::WebsocketClientSend, NotCancellable>::start(
-            self,
-            HostRequestWebsocketSend {
-                message: message_to_serializable(&message),
-            },
-            DurableFunctionType::WriteRemote,
-        )
-        .await?;
+        let mut call =
+            DurableCallSession::<host_functions::WebsocketClientSend, NotCancellable>::start(
+                self,
+                HostRequestWebsocketSend {
+                    message: message_to_serializable(&message),
+                },
+                DurableFunctionType::WriteRemote,
+            )
+            .await?;
 
         if !call.is_live() {
             let _ = self.as_wasi_view().table().get(&self_)?;
@@ -346,15 +348,16 @@ impl<Ctx: WorkerCtx> HostWebsocketConnection for DurableWorkerCtx<Ctx> {
     ) -> anyhow::Result<Result<(), Error>> {
         self.observe_function_call("golem:websocket/client", "close");
 
-        let mut call = CallHandle::<host_functions::WebsocketClientClose, NotCancellable>::start(
-            self,
-            HostRequestWebsocketClose {
-                code,
-                reason: reason.clone(),
-            },
-            DurableFunctionType::WriteRemote,
-        )
-        .await?;
+        let mut call =
+            DurableCallSession::<host_functions::WebsocketClientClose, NotCancellable>::start(
+                self,
+                HostRequestWebsocketClose {
+                    code,
+                    reason: reason.clone(),
+                },
+                DurableFunctionType::WriteRemote,
+            )
+            .await?;
 
         let terminal_close_error =
             TerminalWebSocketError::Closed(Some(SerializableWebsocketCloseInfo {
@@ -492,7 +495,7 @@ impl<U: Send + 'static, Ctx: WorkerCtx> HostWebsocketConnectionWithStore<U>
                 .observe_function_call("golem:websocket/client", "receive")
         });
 
-        let mut call = CallHandle::<
+        let mut call = DurableCallSession::<
             host_functions::WebsocketClientReceive,
             LeaveIncompleteOnDrop,
         >::start_access(
@@ -609,7 +612,7 @@ impl<U: Send + 'static, Ctx: WorkerCtx> HostWebsocketConnectionWithStore<U>
                 .observe_function_call("golem:websocket/client", "receive-with-timeout")
         });
 
-        let mut call = CallHandle::<
+        let mut call = DurableCallSession::<
             host_functions::WebsocketClientReceiveWithTimeout,
             LeaveIncompleteOnDrop,
         >::start_access(
@@ -769,7 +772,7 @@ async fn ensure_websocket_connection_live<Ctx: WorkerCtx>(
 
     // The read-only side-effect trap fires earlier: every caller of this helper
     // (`send` / `receive` / `receive-with-timeout` / `close`) goes through
-    // `CallHandle::start` with `WriteRemote` first, which routes through
+    // `DurableCallSession::start` with `WriteRemote` first, which routes through
     // `DurabilityHost::begin_durable_function` — the single central read-only guard.
     let request = match build_request(&info.url, info.headers.as_deref()) {
         Ok(request) => request,
@@ -871,7 +874,7 @@ async fn ensure_websocket_connection_live_access<U: Send + 'static, Ctx: WorkerC
     }
 
     // The read-only side-effect trap fires earlier: every caller of this helper goes through
-    // `CallHandle::start_access` with `WriteRemote` first, which routes through
+    // `DurableCallSession::start_access` with `WriteRemote` first, which routes through
     // `DurabilityHost::begin_durable_function` — the single central read-only guard.
     let request = match build_request(&info.url, info.headers.as_deref()) {
         Ok(request) => request,
